v1.1.1 - streamlining indexing and pipeline ease

+++ command_line/gen.py
@@ -117,6 +117,7 @@ def matrixForFilename(filename, output):
 	print "Image", path, "has", len(experiments), "experiments."
 	printExperiments(experiments, filename, output)
 
+
 """
 Goes through each image path provided and searches for an associated
 _*strong.list file and _*experiments.json file. Prints relevant data
@@ -127,16 +128,21 @@ arg: path to an individual file.
 def printData(path):
 	global output
 	baseFile = baseName(path)
-	spotListName = "_" + baseFile + "_strong.list"
+	spotPickleName = "_" + baseFile + "_strong.pickle"
+	spotListName = spotPickleName.replace('pickle', 'list')
 	experimentsJson = "_" + baseFile + "_experiments.json"
+	from cppxfel import dials_print_spots
 	
 	print "Finding data for", path
 	
 	if (os.path.isfile(path)):
-		print >> output, "image", path
+		print >> output, "image", baseFile
 	
-	if os.path.isfile(spotListName):
+	if os.path.isfile(spotPickleName):
 		print "Found spots."
+		dials_print_spots.printSpots([spotPickleName])
+	
+	if (os.path.isfile(spotListName)):
 		print >> output, "spots", spotListName
 	else:
 		print >> output, "spots find"
@@ -146,7 +152,9 @@ def printData(path):
 		matrixForFilename(experimentsJson, output)
 
 def dumpPanels(images):
+	import pickle
 	global panels
+	global distance, wavelength, centre, pixelSize
 	print "Attempting to dump panel file"
 	# i is an integer specifying the image counter
 	i = 0
@@ -168,7 +176,19 @@ def dumpPanels(images):
 		# now we dump from this pickle file and stop the while loop
 		
 		successful = True
-		panels = x['ACTIVE_AREAS']
+		
+		try:
+			panels = x['ACTIVE_AREAS']
+			distance = x['DISTANCE']
+			wavelength = x['WAVELENGTH']
+			pixelSize = (x['PIXEL_SIZE'], 0)
+			centre = (x['BEAM_CENTER_X'] / pixelSize[0], x['BEAM_CENTER_Y'] / pixelSize[0])
+		except Exception as e:
+			print Exception
+			break
+		
+		# TODO: find distance, wavelength, beam centre from first file.
+		
 		panelTxt = StringIO.StringIO()
 	
 		for i in range(0, len(panels), 4):
@@ -193,10 +213,22 @@ def dumpPanels(images):
 	file.close()
 	return
 
+"""
+	Dumps images specified in the imagePath list from
+	pickle form into img form, but will skip those
+	which have already been dumped.
+	arg imagePaths: list of strings.
+"""
 def dumpImages(imagePaths):
 	for path in imagePaths:
 		print "Dumping image:", path
 		rootname = splitext(basename(path))[0]
+		imageName = rootname + '.img'
+		
+		if (os.path.exists(imageName)):
+			print "Image", imageName, "already dumped - skipping."
+			continue
+		
 		ext = splitext(basename(path))[1]
 		
 		if not ext == ".pickle":
@@ -213,9 +245,13 @@ def dumpImages(imagePaths):
 			data_array.append(data[i])
 
 		string = data_array.tostring()
-		newFile = open(rootname + '.img', 'wb')
+		newFile = open(imageName, 'wb')
 		newFile.write(string)
 		newFile.close()
+
+# **********************************
+# ***** START OF EXECUTED CODE *****
+# **********************************
 
 # Take all the paths provided after cppxfel.input_gen.
 paths = sys.argv[1:]
@@ -223,7 +259,7 @@ paths = sys.argv[1:]
 if (len(paths) == 0):
 	print "Please specify the images you want to dump, e.g.:"
 	print "\tcppxfel.input_gen shot*.pickle"
-	print "\tcppxfel.input_gen shot*.cbf"
+#	print "\tcppxfel.input_gen shot*.cbf"
 	exit()
 
 # Start the output stringIO which will be written to matrices.dat
@@ -233,9 +269,7 @@ output = StringIO.StringIO()
 for path in paths:
 	printData(path)
 
-print "Contents of matrices.dat file:"
-print output.getvalue()
-print "End of matrices.dat file contents."
+print "Written data to matrices.dat file."
 
 outputFilename = "matrices.dat"
 outputFile = open(outputFilename, 'w')
@@ -270,18 +304,78 @@ for i in range(0, thread_count):
 
 print "Creating input files"
 
+indexTxt = StringIO.StringIO()
+
+print >> indexTxt, "ORIENTATION_MATRIX_LIST matrices.dat"
+print >> indexTxt, "NEW_MATRIX_LIST indexed.dat"
+
+print >> indexTxt, "# Be sure to set the UNIT_CELL and SPACE_GROUP for indexing. cppxfel cannot index without this knowledge."
+print >> indexTxt, "SPACE_GROUP", spacegroup
+
+if len(unit_cell_dimensions):
+	print >> indexTxt, "UNIT_CELL",
+	for dimension in unit_cell_dimensions:
+		print >> indexTxt, dimension,
+print >> indexTxt, "\n"
+
+print >> indexTxt, "MM_PER_PIXEL", pixelSize[0]
+print >> indexTxt, "BEAM_CENTRE", centre[0], centre[1]
+print >> indexTxt, "DETECTOR_DISTANCE", distance
+print >> indexTxt, "INTEGRATION_WAVELENGTH", wavelength, "\n"
+print >> indexTxt, "PANEL_LIST panels.txt"
+print >> indexTxt, "METROLOGY_SEARCH_SIZE 2\n"
+print >> indexTxt, "# If your crystal is highly mosaic or the detector is quite far back you may need to increase the padding values."
+print >> indexTxt, "SHOEBOX_FOREGROUND_PADDING 1"
+print >> indexTxt, "SHOEBOX_NEITHER_PADDING 2"
+print >> indexTxt, "SHOEBOX_BACKGROUND_PADDING 3\n"
+print >> indexTxt, "# If you see too many spots, increase the intensity threshold.\nINTENSITY_THRESHOLD 12"
+print >> indexTxt, "ABSOLUTE_INTENSITY OFF\n"
+print >> indexTxt, "OVER_PRED_BANDWIDTH 0.07\n"
+print >> indexTxt, "REFINE_ORIENTATIONS ON"
+print >> indexTxt, "ROUGH_CALCULATION ON\n"
+
+print >> indexTxt, "# Specifies maximum multiple lattices to index in total"
+print >> indexTxt, "SOLUTION_ATTEMPTS 1\n"
+
+print >> indexTxt, "# Maximum reciprocal distance from spot to spot to consider for analysis."
+print >> indexTxt, "# A maximum reciprocal distance of 0.1 would be equivalent separation"
+print >> indexTxt, "# between the beam centre and the 10 Angstrom resolution ring."
+print >> indexTxt, "MAX_RECIPROCAL_DISTANCE 0.15\n"
+
+print >> indexTxt, "# Initial rlp size: used to determine the tolerances for the vector lengths in the crystal."
+print >> indexTxt, "# For a 1 micron crystal with no mosaicity, the initial rlp size is 0.0001 Ang^-1 (i.e.,"
+print >> indexTxt, "# 1 / 10000 Ang). To be more strict for indexing, lower this number; to be less strict increase it."
+print >> indexTxt, "INITIAL_RLP_SIZE 0.0001\n"
+
+print >> indexTxt, "# If you wish to see more verbose output, change to 1 (moderate), or 2 (debug, usually too much)."
+print >> indexTxt, "VERBOSITY_LEVEL 0\n"
+print >> indexTxt, "COMMANDS\n"
+print >> indexTxt, "INDEX"
+
+outputFilename = "index.txt"
+
+if not os.path.isfile(outputFilename):
+	outputFile = open(outputFilename, 'w')
+	print >>outputFile, indexTxt.getvalue()
+	outputFile.close()
+	print "New template input file index.txt"
+	print "Please check your target unit cell and space group"
+	
+
 integrateTxt = StringIO.StringIO()
 
 print >> integrateTxt, "ORIENTATION_MATRIX_LIST matrices.dat"
+print >> integrateTxt, "# Enable this line to take results from cppxfel indexing (as opposed to DIALS):"
+print >> integrateTxt, "# ORIENTATION_MATRIX_LIST integrate-indexed.dat"
+
 print >> integrateTxt, "NEW_MATRIX_LIST orientations.dat\n"
 
-if spacegroup > 0:
-	print >> integrateTxt, "SPACE_GROUP", spacegroup
+print >> indexTxt, "# Be sure to set the SPACE_GROUP for integration."
+print >> integrateTxt, "SPACE_GROUP", spacegroup
 
-if len(unit_cell_dimensions):
-	print >> integrateTxt, "UNIT_CELL",
-	for dimension in unit_cell_dimensions:
-		print >> integrateTxt, dimension,
+print >> integrateTxt, "UNIT_CELL",
+for dimension in unit_cell_dimensions:
+	print >> integrateTxt, dimension,
 print >> integrateTxt 
 print >> integrateTxt, "FIX_UNIT_CELL ON\n"
 print >> integrateTxt, "MM_PER_PIXEL", pixelSize[0]
@@ -290,24 +384,29 @@ print >> integrateTxt, "DETECTOR_DISTANCE", distance
 print >> integrateTxt, "INTEGRATION_WAVELENGTH", wavelength, "\n"
 print >> integrateTxt, "PANEL_LIST panels.txt"
 print >> integrateTxt, "METROLOGY_SEARCH_SIZE 1\n"
+print >> integrateTxt, "# If your crystal is highly mosaic or the detector is quite far back you may need to increase the padding values."
 print >> integrateTxt, "SHOEBOX_FOREGROUND_PADDING 1"
 print >> integrateTxt, "SHOEBOX_NEITHER_PADDING 2"
 print >> integrateTxt, "SHOEBOX_BACKGROUND_PADDING 3\n"
+print >> integrateTxt, "# If you see too many spots, increase the intensity threshold."
 print >> integrateTxt, "INTENSITY_THRESHOLD 12"
-print >> integrateTxt, "ABSOLUTE_INTENSITY OFF"
+print >> integrateTxt, "ABSOLUTE_INTENSITY OFF\n"
 print >> integrateTxt, "REFINE_ORIENTATIONS ON\n"
-print >> integrateTxt, "VERBOSITY_LEVEL 1\n"
+print >> integrateTxt, "# Enable the following line to be a rougher, quicker calculation"
+print >> integrateTxt, "# ROUGH_CALCULATION ON\n"
+print >> integrateTxt, "VERBOSITY_LEVEL 0\n"
 print >> integrateTxt, "COMMANDS\n"
 print >> integrateTxt, "INTEGRATE"
 
 outputFilename = "integrate.txt"
-outputFile = open(outputFilename, 'w')
-print >>outputFile, integrateTxt.getvalue()
-outputFile.close()
 
-print "New template input file integrate.txt"
-print "Please check your target unit cell and space group"
-
+if not os.path.isfile(outputFilename):
+	outputFile = open(outputFilename, 'w')
+	print >>outputFile, integrateTxt.getvalue()
+	outputFile.close()
+	print "New template input file integrate.txt"
+	print "Please check your target unit cell and space group"
+	
 refineTxt = StringIO.StringIO()
 
 print >> refineTxt, "ORIENTATION_MATRIX_LIST refine-orientations.dat"
@@ -318,11 +417,13 @@ print >> refineTxt, "COMMANDS\n"
 print >> refineTxt, "REFINE_PARTIALITY"
 
 outputFilename = "refine.txt"
-outputFile = open(outputFilename, 'w')
-print >>outputFile, refineTxt.getvalue()
-outputFile.close()
 
-print "New template input file refine.txt"
+if not os.path.isfile(outputFilename):
+	outputFile = open(outputFilename, 'w')
+	print >>outputFile, refineTxt.getvalue()
+	outputFile.close()
+	print "New template input file refine.txt"
+
 
 mergeTxt = StringIO.StringIO()
 
@@ -337,11 +438,37 @@ print >> mergeTxt, "COMMANDS\n"
 print >> mergeTxt, "MERGE"
 
 outputFilename = "merge.txt"
-outputFile = open(outputFilename, 'w')
-print >>outputFile, mergeTxt.getvalue()
-outputFile.close()
 
-print "New template input file merge.txt"
+if not os.path.isfile(outputFilename):
+	outputFile = open(outputFilename, 'w')
+	print >>outputFile, mergeTxt.getvalue()
+	outputFile.close()
+	print "New template input file merge.txt"
 
 if distance == 0:
-	print "You MUST change the experimental parameters in integrate.txt"
+	print "You MUST change the experimental parameters in index.txt and integrate.txt"
+
+
+print "Now what?"
+print "********************"
+print "***** INDEXING *****"
+print "********************"
+print
+print "To index, change the space group and unit cell dimensions in index.txt."
+print "After indexing, alter integrate.txt to accept the alternative value for"
+print "ORIENTATION_MATRIX_LIST (uncomment the line)."
+print
+print "Command to index: cppxfel.run -i index.txt"
+print
+print "********************"
+print "*** INTEGRATION ****"
+print "********************"
+print
+print "To integrate orientations from indexing using DIALS, simply skip index.txt."
+print "CHECK the unit cell and space group listed in integrate.txt, and then"
+print "run the integrate.txt input file. This includes initial orientation matrix"
+print "refinement."
+print
+print "Command to integrate: cppxfel.run -i integrate.txt"
+print
+print "... but first, wait for image dumps:"
